+++ pkg/translator/kubemapper.go
@@ -7,13 +7,13 @@ import (
 	"strings"
 
 	log "github.com/Sirupsen/logrus"
-
 	// metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
-
 	"github.com/pearsontechnology/environment-operator/pkg/bitesize"
+	"github.com/pearsontechnology/environment-operator/pkg/config"
 	ext "github.com/pearsontechnology/environment-operator/pkg/k8_extensions"
 	"github.com/pearsontechnology/environment-operator/pkg/util"
 	// metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
+	"github.com/pearsontechnology/environment-operator/pkg/util/k8s"
 	"k8s.io/client-go/pkg/api/resource"
 	"k8s.io/client-go/pkg/api/unversioned"
 	"k8s.io/client-go/pkg/api/v1"
@@ -143,6 +143,7 @@ func (w *KubeMapper) Deployment() (*v1beta1.Deployment, error) {
 						"name":        w.BiteService.Name,
 						"version":     w.BiteService.Version,
 					},
+					Annotations: w.BiteService.Annotations,
 				},
 				Spec: v1.PodSpec{
 					NodeSelector:     map[string]string{"role": "minion"},
@@ -239,6 +240,9 @@ func (w *KubeMapper) container() (*v1.Container, error) {
 
 func (w *KubeMapper) envVars() ([]v1.EnvVar, error) {
 	var retval []v1.EnvVar
+	var err error
+	//Create in cluster rest client to be utilized for secrets processing
+	client, _ := k8s.ClientForNamespace(config.Env.Namespace)
 
 	for _, e := range w.BiteService.EnvVars {
 		var evar v1.EnvVar
@@ -253,6 +257,11 @@ func (w *KubeMapper) envVars() ([]v1.EnvVar, error) {
 			} else {
 				secretName = kv[0]
 				secretDataKey = secretName
+			}
+
+			if !client.Secret().Exists(secretName) {
+				log.Debugf("Unable to Find Secret %s", secretName)
+				err = fmt.Errorf("Unable to find secret [%s] in namespace [%s] when processing envvars for deployment [%s]", secretName, config.Env.Namespace, w.BiteService.Name)
 			}
 
 			evar = v1.EnvVar{
@@ -274,21 +283,7 @@ func (w *KubeMapper) envVars() ([]v1.EnvVar, error) {
 		}
 		retval = append(retval, evar)
 	}
-	return retval, nil
-}
-
-func (w *KubeMapper) Annotations() ([]v1.ObjectMeta, error) {
-	var retval []v1.ObjectMeta
-
-	for _, a := range w.BiteService.Annotations {
-		annotation := v1.ObjectMeta {
-			Annotations: map[string]string{
-				a.Name: a.Value,
-			},
-		}
-		retval = append(retval, annotation)
-	}
-	return retval, nil
+	return retval, err
 }
 
 func (w *KubeMapper) volumeMounts() ([]v1.VolumeMount, error) {